+++ CHANGELOG.rst
@@ -1,9 +1,10 @@
 Changelog
 =========
 
-Current (2024-05-28)
+Current (2024-05-29)
 --------------------
 
+* Implemented the i2c_connection_helper base class
 * Added word_list_to_bytes and bytes_to_word_list functions
 * Added address byte adjustment function for different endianness and bit lengths
 * Added I2C address validation function